+++ test/integration/connection/test-typecast.test.cjs
@@ -9,15 +9,15 @@ const connection = common.createConnection();
 connection.query('CREATE TEMPORARY TABLE json_test (json_test JSON)');
 connection.query(
   'INSERT INTO json_test VALUES (?)',
-  JSON.stringify({ test: 42 }),
+  JSON.stringify({ test: 42 })
 );
 
 connection.query(
-  'CREATE TEMPORARY TABLE geom_test (p POINT, g GEOMETRY NOT NULL)',
+  'CREATE TEMPORARY TABLE geom_test (p POINT, g GEOMETRY NOT NULL)'
 );
 connection.query(
   'INSERT INTO geom_test VALUES (ST_GeomFromText("POINT(1 1)"), ' +
-    'ST_GeomFromText("LINESTRING(-71.160281 42.258729,-71.160837 42.259113,-71.161144 42.25932)"))',
+    'ST_GeomFromText("LINESTRING(-71.160281 42.258729,-71.160837 42.259113,-71.161144 42.25932)"))'
 );
 
 connection.query(
@@ -34,7 +34,7 @@ connection.query(
   (err, res) => {
     assert.ifError(err);
     assert.equal(res[0].foo, 'FOO UPPERCASE');
-  },
+  }
 );
 
 connection.query(
@@ -44,9 +44,9 @@ connection.query(
   },
   (err, res) => {
     assert.ifError(err);
-    assert(Buffer.isBuffer(res[0].foo));
+    assert(Buffer.isBuffer(res[0].foo), 'Check for Buffer');
     assert.equal(res[0].foo.toString('utf8'), 'foobar');
-  },
+  }
 );
 
 connection.query(
@@ -60,7 +60,7 @@ connection.query(
     assert.ifError(err);
     assert.equal(_rows[0].test, null);
     assert.equal(_rows[0].value, 6);
-  },
+  }
 );
 
 connection.query(
@@ -73,7 +73,7 @@ connection.query(
   (err, _rows) => {
     assert.ifError(err);
     assert.equal(_rows[0].json_test.test, 42);
-  },
+  }
 );
 
 connection.execute(
@@ -86,7 +86,7 @@ connection.execute(
   (err, _rows) => {
     assert.ifError(err);
     assert.equal(_rows[0].json_test.test, 42);
-  },
+  }
 );
 
 // read geo fields
@@ -103,9 +103,9 @@ connection.query(
         { x: -71.160837, y: 42.259113 },
         { x: -71.161144, y: 42.25932 },
       ],
-      res[0].g,
+      res[0].g
     );
-  },
+  }
 );
 
 connection.query(
@@ -126,7 +126,7 @@ connection.query(
             { x: -71.160837, y: 42.259113 },
             { x: -71.161144, y: 42.25932 },
           ],
-          field.geometry(),
+          field.geometry()
         );
 
         return [
@@ -150,9 +150,9 @@ connection.query(
         { x: -60, y: 50 },
         { x: -50, y: 60 },
       ],
-      res[0].g,
+      res[0].g
     );
-  },
+  }
 );
 
 connection.end();