+++ typings/mysql/lib/Connection.d.ts
@@ -323,6 +323,8 @@ export interface ConnectionOptions {
 
   waitForConnections?: boolean;
 
+  disableEval?: boolean;
+
   authPlugins?: {
     [key: string]: AuthPlugin;
   };
@@ -353,11 +355,7 @@ declare class BaseConnection extends QueryableBase(
       | ResultSetHeader,
   >(
     sql: string,
-    callback?: (
-      err: QueryError | null,
-      result: T,
-      fields: FieldPacket[],
-    ) => any,
+    callback?: (err: QueryError | null, result: T, fields: FieldPacket[]) => any
   ): Query;
   static createQuery<
     T extends
@@ -369,11 +367,7 @@ declare class BaseConnection extends QueryableBase(
   >(
     sql: string,
     values: any | any[] | { [param: string]: any },
-    callback?: (
-      err: QueryError | null,
-      result: T,
-      fields: FieldPacket[],
-    ) => any,
+    callback?: (err: QueryError | null, result: T, fields: FieldPacket[]) => any
   ): Query;
 
   beginTransaction(callback: (err: QueryError | null) => void): void;
@@ -384,7 +378,7 @@ declare class BaseConnection extends QueryableBase(
 
   changeUser(
     options: ConnectionOptions,
-    callback?: (err: QueryError | null) => void,
+    callback?: (err: QueryError | null) => void
   ): void;
 
   end(callback?: (err: QueryError | null) => void): void;
@@ -409,7 +403,7 @@ declare class BaseConnection extends QueryableBase(
 
   prepare(
     sql: string,
-    callback?: (err: QueryError | null, statement: PrepareStatementInfo) => any,
+    callback?: (err: QueryError | null, statement: PrepareStatementInfo) => any
   ): Prepare;
 
   unprepare(sql: string): PrepareStatementInfo;