+++ typings/mysql/lib/protocol/sequences/Prepare.d.ts
@@ -19,11 +19,7 @@ export class PrepareStatementInfo {
       | ResultSetHeader,
   >(
     parameters: any | any[] | { [param: string]: any },
-    callback?: (
-      err: QueryError | null,
-      result: T,
-      fields: FieldPacket[],
-    ) => any,
+    callback?: (err: QueryError | null, result: T, fields: FieldPacket[]) => any
   ): Query;
 }
 
@@ -57,11 +53,11 @@ declare class Prepare extends Sequence {
   on(event: 'error', listener: (err: QueryError) => any): this;
   on(
     event: 'fields',
-    listener: (fields: FieldPacket, index: number) => any,
+    listener: (fields: FieldPacket, index: number) => any
   ): this;
   on(
     event: 'result',
-    listener: (result: RowDataPacket | OkPacket, index: number) => any,
+    listener: (result: RowDataPacket | OkPacket, index: number) => any
   ): this;
   on(event: 'end', listener: () => any): this;
 }