+++ typings/mysql/lib/protocol/sequences/QueryableBase.d.ts
@@ -7,34 +7,34 @@ import {
 } from './Query.js';
 
 export declare function QueryableBase<T extends QueryableConstructor>(
-  Base?: T,
+  Base?: T
 ): {
   new (...args: any[]): {
     query<T extends QueryResult>(
       sql: string,
       callback?:
         | ((err: QueryError | null, result: T, fields: FieldPacket[]) => any)
-        | undefined,
+        | undefined
     ): Query;
     query<T extends QueryResult>(
       sql: string,
       values: any,
       callback?:
         | ((err: QueryError | null, result: T, fields: FieldPacket[]) => any)
-        | undefined,
+        | undefined
     ): Query;
     query<T extends QueryResult>(
       options: QueryOptions,
       callback?:
         | ((err: QueryError | null, result: T, fields?: FieldPacket[]) => any)
-        | undefined,
+        | undefined
     ): Query;
     query<T extends QueryResult>(
       options: QueryOptions,
       values: any,
       callback?:
         | ((err: QueryError | null, result: T, fields: FieldPacket[]) => any)
-        | undefined,
+        | undefined
     ): Query;
   };
 } & T;