+++ website/src/components/Stability.tsx
@@ -1,4 +1,4 @@
-import { FC, JSX } from 'react';
+import type { FC, JSX } from 'react';
 import Link from '@docusaurus/Link';
 import {
   AlertTriangle,
@@ -45,11 +45,9 @@ export type StabilityProps = {
  * ```
  */
 export const Stability: FC<StabilityProps> = ({ level, message }) => {
-  /* eslint-disable @typescript-eslint/indent */
   const styles: Record<
     StabilityProps['level'],
     { title: string; icon: JSX.Element }
-    /* eslint-enable @typescript-eslint/indent */
   > = {
     0: {
       title: 'Deprecated',